requirements.txt
streamlit==1.31.1
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
pymongo
bcrypt
requests
langchain==0.1.4
langchain_community==0.0.16
langchain_core==0.1.17
google-generativeai==0.5.4
python-dotenv==1.0.1
pillow